写完啦，一共经历了两次大bug修改

+++ service/src/main/java/cc/ccocc/service/IUser_ArchivesService.java
@@ -1,7 +1,5 @@
 package cc.ccocc.service;
 
-import org.apache.ibatis.annotations.Param;
-
 /**
  * Created on 21:42  03/02/2020
  * Description:
@@ -10,5 +8,5 @@ import org.apache.ibatis.annotations.Param;
  */
 public interface IUser_ArchivesService {
 
-    Integer addInUser_Archives( Long userId,  int archiveId );
+    Integer addInUser_Archives( Long userId,  Long archiveId );
 }
